Adding supermatrix_factory, which computes Jacobians of functions.

+++ core/point.h
@@ -180,7 +180,7 @@ static inline void point_randomize(point_t* point, int (*rand_gen)(), bbox_t* bo
 // Given a random number generator, generate a vector with the given magnitude
 // pointing in a random direction. The random number generator must generate 
 // an integer between 0 and RAND_MAX.
-static inline void vector_randomize(vector_t* vector, long (*rand_gen)(), double magnitude)
+static inline void vector_randomize(vector_t* vector, int (*rand_gen)(), double magnitude)
 {
   double theta = (1.0*rand_gen()/RAND_MAX) * M_PI;
   double phi = (1.0*rand_gen()/RAND_MAX) * 2.0 * M_PI;
